Clarifying the manifest file change location
Hi @Shin-NiL, I just wanted to add that line to Clarify a bit more on the readme, where most people follow, where
those three lines are. Because I made the mistake of putting it in the wrong location in AndroidManifest.xml file at first and then went to the link and found out my mistake. So this could help a bit more at first sight.